Fix prereqTask dialog & add AutoComplete search
